requirements.txt
astroid==1.6.1
awacs==0.7.2
awscli==1.14.53
botocore==1.9.6
cfn-flip==1.0.1
click==6.7
colorama==0.3.7
docutils==0.14
isort==4.3.4
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pyasn1==0.4.2
pylint==1.8.2
python-dateutil==2.6.1
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
troposphere==2.2.0
wrapt==1.10.11
